new directories before touching the controller
