+++ src/main/java/org/Yaed/windows/cuartosM/HabitacionM2.java
@@ -1,13 +1,13 @@
-package org.Yaed.windows;
+package org.Yaed.windows.cuartosM;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableCellRenderer;
 import javax.swing.table.DefaultTableModel;
 import java.awt.*;
 
-public class HabitacionF8 extends JFrame {
-    public HabitacionF8() {
-        setTitle("Casa Pequeña 8");
+public class HabitacionM2 extends JFrame {
+    public HabitacionM2() {
+        setTitle("Casa Mediana 2");
         setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
         setSize(700, 500);
         setLocationRelativeTo(null);
@@ -23,16 +23,19 @@ public class HabitacionF8 extends JFrame {
         mainPanel.setLayout(new BorderLayout(20, 20));
         mainPanel.setBorder(BorderFactory.createEmptyBorder(20, 30, 20, 30));
 
-        JLabel titleLabel = new JLabel("Casa Pequeña 8 (Femenino)");
+        // Título
+        JLabel titleLabel = new JLabel("Casa Mediana 2 (Masculino)");
         titleLabel.setFont(new Font("Segoe UI", Font.BOLD, 26));
         titleLabel.setForeground(accentColor);
         titleLabel.setHorizontalAlignment(SwingConstants.CENTER);
 
-        JLabel subtitleLabel = new JLabel("Cuarto 1");
+        // Subtítulo
+        JLabel subtitleLabel = new JLabel("Cuarto 2");
         subtitleLabel.setFont(new Font("Segoe UI", Font.PLAIN, 16));
         subtitleLabel.setForeground(Color.DARK_GRAY);
         subtitleLabel.setHorizontalAlignment(SwingConstants.CENTER);
 
+        // Selector de mes y año (reemplaza el botón buscar por un combo de año)
         JPanel selectorPanel = new JPanel();
         selectorPanel.setBackground(bgColor);
         selectorPanel.setLayout(new FlowLayout(FlowLayout.CENTER, 10, 0));
@@ -60,10 +63,12 @@ public class HabitacionF8 extends JFrame {
         tabla.setSelectionBackground(new Color(204, 229, 255));
         tabla.setSelectionForeground(Color.BLACK);
 
+        // Encabezado bonito
         tabla.getTableHeader().setFont(new Font("Segoe UI", Font.BOLD, 15));
         tabla.getTableHeader().setBackground(tableHeaderColor);
         tabla.getTableHeader().setForeground(accentColor);
 
+        // Centrar columnas
         DefaultTableCellRenderer centerRenderer = new DefaultTableCellRenderer();
         centerRenderer.setHorizontalAlignment(SwingConstants.CENTER);
         for (int i = 0; i < tabla.getColumnCount(); i++) {
@@ -73,6 +78,13 @@ public class HabitacionF8 extends JFrame {
         JScrollPane scrollPane = new JScrollPane(tabla);
         scrollPane.setBorder(BorderFactory.createLineBorder(accentColor, 1));
 
+        // Layout
+        mainPanel.add(titleLabel, BorderLayout.NORTH);
+        mainPanel.add(subtitleLabel, BorderLayout.BEFORE_FIRST_LINE);
+        mainPanel.add(selectorPanel, BorderLayout.CENTER);
+        mainPanel.add(scrollPane, BorderLayout.SOUTH);
+
+        // Ajuste de layout para que todo quede bien distribuido
         JPanel topPanel = new JPanel(new GridLayout(2, 1));
         topPanel.setBackground(bgColor);
         topPanel.add(titleLabel);
