+++ src/main/java/org/Yaed/windows/forms/FormularioCarrera.java
@@ -1,4 +1,4 @@
-package org.Yaed.windows;
+package org.Yaed.windows.forms;
 
 import org.Yaed.controller.CarrerasController;
 import org.Yaed.entity.Carrera;
